Day 14: make use of Long instead of Integer because of Part 2 input

+++ src/main/java/nl/smerik/adventofcode/aoc2019/model/nanofactory/ChemicalUnits.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class ChemicalUnits {
 
     private final String chemical;
-    private final int units;
+    private final long units;
 }
